Renamed bloc directory to application
