+++ lib/bloc.dart
@@ -1,4 +1,4 @@
 /// Contains the bloc related files.
 library bloc;
 
-export 'src/bloc/b_router_cubit.dart';
+export 'src/application/b_router_cubit.dart';
