+++ lib/src/router/parser.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-import '../bloc/b_router_cubit.dart';
+import '../application/b_router_cubit.dart';
 import 'route.dart';
 
 /// The [RouteInformationParser] implementation for our [MaterialApp.router]
